Extract SpectralEmbedding to feature_extraction module
