+++ divik/__init__.py
@@ -3,6 +3,7 @@ __version__ = '2.3.5'
 from ._seeding import seeded
 from ._utils import DivikResult
 from divik import feature_selection
+from divik import feature_extraction
 from divik import cluster
 from ._summary import plot, reject_split
 
@@ -10,6 +11,7 @@ __all__ = [
     "__version__",
     "cluster",
     "feature_selection",
+    "feature_extraction",
     "seeded",
     'DivikResult',
     "plot", "reject_split",
