+++ divik/feature_extraction/__init__.py
@@ -1,0 +1,5 @@
+from ._spectral import LocallyAdjustedRbfSpectralEmbedding
+
+__all__ = [
+    'LocallyAdjustedRbfSpectralEmbedding',
+]
