+++ docs/modules/divik.feature_extraction.rst
@@ -1,0 +1,7 @@
+`feature\_extraction` module
+===========================
+
+.. automodule:: divik.feature_extraction
+   :members:
+   :undoc-members:
+   :inherited-members:
